# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeIt API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StockQuote {
  ticker: string;
  companyName: string;
  price: number;
  change: number;
  changePercent: number;
  marketCap?: string;
  volume?: string;
  description: string;
  sector?: string;
  high52w?: number;
  low52w?: number;
}

export interface TrendingStock {
  ticker: string;
  companyName: string;
  price: number;
  changePercent: number;
}

export interface TrendingStocksResponse {
  stocks: TrendingStock[];
}

export interface ApiError {
  error: string;
}

export type GetStockQuoteParams = {
  ticker: string;
};