# lib/api-zod/src/generated/types/stockQuote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeIt API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StockQuote {
  ticker: string;
  companyName: string;
  price: number;
  change: number;
  changePercent: number;
  marketCap?: string;
  volume?: string;
  description: string;
  sector?: string;
  high52w?: number;
  low52w?: number;
}