# lib/api-zod/src/generated/types/trendingStock.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeIt API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrendingStock {
  ticker: string;
  companyName: string;
  price: number;
  changePercent: number;
}