# lib/api-zod/src/generated/types/trendingStocksResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeIt API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrendingStock } from "./trendingStock";

export interface TrendingStocksResponse {
  stocks: TrendingStock[];
}